+++ plutarch-ledger-api/CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
     * `PScriptInfo`
     * `PCommittee`
 * `PRationalData`, originally from `plutarch-extra`
+* `PTryFrom PData (PAsData PLovelace)` instance
 
 ### Modified